updated pages,status

+++ src/actions/tasks.ts
@@ -1,17 +1,31 @@
-import { PrismaClient } from '@prisma/client';
-import { z } from 'zod';
+import { AstroCookies } from 'astro';
 //import { v4 as uuidv4 } from 'uuid';
-import { defineAction, ActionError } from 'astro:actions';
+import { defineAction } from 'astro:actions';
+import { type ActionAPIContext } from 'astro:actions';
+import { z } from 'zod';
+import { type TaskInput, TaskUpdate } from '../utils/types';
+import prisma from '../prisma';
 
-const prisma = new PrismaClient();
+interface Action {
+  handler: (input: any, context: any) => Promise<any>;
+}
 
+const checkCookie = (cookies: AstroCookies): string | null => {
+  // Чтение userId из cookies
+  const userCookie = cookies.get('userId');
+  const userId = userCookie ? userCookie.value : null;
+
+  if (!userId) {
+    throw new Error('Не удалось получить задачи, так как идентификатор пользователя не найден в куках.');
+  }
+  return userId;
+};
 // Определяем схему данных для валидации
-const taskSchema = z.object({
+const taskSchemaInput = z.object({
   title: z.string().min(1, 'Название задачи не может быть пустым'),
   description: z.string().optional(),
   priority: z.enum(['low', 'medium', 'high']),
   dueDate: z.string().optional(),
-  status: z.enum(['active', 'completed']).default('active'),
 });
 
 const taskUpdateSchema = z.object({
@@ -24,14 +38,8 @@ const taskUpdateSchema = z.object({
 });
 
 export const getTasks = defineAction({
-  handler: async (input, { cookies }) => {
-    // Чтение userId из cookies
-    const userCookie = cookies.get('userId');
-    const userId = userCookie ? userCookie.value : null;
-
-    if (!userId) {
-      throw new Error('Не удалось получить задачи, так как идентификатор пользователя не найден в куках.');
-    }
+  handler: async (input: null, { cookies }: any) => {
+    const userId = checkCookie(cookies);
 
     // Получение задач только для конкретного пользователя
     const tasks = await prisma.task.findMany({
@@ -46,15 +54,9 @@ export const getTasks = defineAction({
 
 export const createTask = defineAction({
   accept: 'form',
-  input: taskSchema,
-  handler: async (input, { cookies }) => {
-    // Чтение userId из cookies
-    const userCookie = cookies.get('userId');
-    const userId = userCookie ? userCookie.value : null;
-
-    if (!userId) {
-      throw new Error('Не удалось создать задачу, так как идентификатор пользователя не найден в куках.');
-    }
+  input: taskSchemaInput,
+  handler: async (input, { cookies }: ActionAPIContext) => {
+    const userId = checkCookie(cookies);
     // Создаем новую запись в базе данных
     const newTask = await prisma.task.create({
       data: {
@@ -65,7 +67,7 @@ export const createTask = defineAction({
         dueDate: input.dueDate ? new Date(input.dueDate) : null, // Преобразуем строку в дату, если передано
       },
     });
-
+    console.log('createTask', newTask);
     return newTask;
   },
 });
@@ -74,7 +76,9 @@ export const getTaskById = defineAction({
   input: z.object({
     id: z.string(),
   }),
-  handler: async (input) => {
+  handler: async (input: { id: string }, { cookies }: { cookies: AstroCookies }) => {
+    const userId = checkCookie(cookies);
+
     const task = await prisma.task.findUnique({
       where: { id: input.id },
     });
@@ -85,7 +89,9 @@ export const getTaskById = defineAction({
 export const updateTaskById = defineAction({
   accept: 'form',
   input: taskUpdateSchema,
-  handler: async (input) => {
+  handler: async (input: TaskUpdate, { cookies }: { cookies: AstroCookies }) => {
+    const userId = checkCookie(cookies);
+
     const updatedTask = await prisma.task.update({
       where: { id: input.id },
       data: {
@@ -104,8 +110,15 @@ export const deleteTaskById = defineAction({
   input: z.object({
     id: z.string(),
   }),
-  handler: async (input) => {
-    console.log('deleteTaskById', input.id);
+  handler: async (input: { id: string }, { cookies }: { cookies: AstroCookies }) => {
+    // Чтение userId из cookies
+    const userCookie = cookies.get('userId');
+    const userId = userCookie ? userCookie.value : null;
+
+    if (!userId) {
+      throw new Error('Не удалось создать задачу, так как идентификатор пользователя не найден в куках.');
+    }
+
     const deletedTask = await prisma.task.delete({
       where: { id: input.id },
     });
